# docs/swagger/swagger.yaml
basePath: /
definitions:
  domain.APIKey:
    properties:
      created_at:
        type: string
      expires_at:
        type: string
      id:
        type: string
      key_prefix:
        type: string
      last_used_at:
        type: string
      name:
        type: string
      revoked_at:
        type: string
      tenant_id:
        type: string
      user_id:
        type: string
    type: object
  domain.Actor:
    properties:
      api_key_id:
        type: string
      email:
        type: string
      is_platform_admin:
        type: boolean
      role:
        $ref: '#/definitions/domain.Role'
      tenant_id:
        type: string
      user_id:
        type: string
    type: object
  domain.AdminInstance:
    properties:
      app:
        $ref: '#/definitions/domain.App'
      deployment:
        $ref: '#/definitions/domain.Deployment'
      user_email:
        type: string
    type: object
  domain.AdminSummary:
    properties:
      apps:
        type: integer
      deployments:
        type: integer
      failed_deployments:
        type: integer
      repository_driver:
        type: string
      tenants:
        type: integer
      users:
        type: integer
    type: object
  domain.App:
    properties:
      config:
        $ref: '#/definitions/domain.AppConfig'
      created_at:
        type: string
      created_by:
        type: string
      current_deployment_id:
        type: string
      desired_state:
        type: string
      id:
        type: string
      name:
        type: string
      slug:
        type: string
      tenant_id:
        type: string
      updated_at:
        type: string
    type: object
  domain.AppConfig:
    properties:
      cpu_limit:
        type: string
      cpu_request:
        type: string
      domain:
        type: string
      env:
        additionalProperties:
          type: string
        type: object
      image:
        type: string
      memory_limit:
        type: string
      memory_request:
        type: string
      openclaw:
        $ref: '#/definitions/domain.OpenClawConfig'
      port:
        type: integer
      public:
        type: boolean
      reliability:
        $ref: '#/definitions/domain.ReliabilityPolicy'
      replicas:
        type: integer
      service_account_name:
        type: string
    type: object
  domain.AuditLog:
    properties:
      action:
        type: string
      actor_api_key_id:
        type: string
      actor_user_id:
        type: string
      created_at:
        type: string
      id:
        type: string
      message:
        type: string
      resource_id:
        type: string
      resource_type:
        type: string
      tenant_id:
        type: string
    type: object
  domain.BackendCapabilities:
    properties:
      capabilities:
        items:
          $ref: '#/definitions/domain.BackendCapability'
        type: array
      display_name:
        type: string
      name:
        type: string
      notes:
        items:
          type: string
        type: array
    type: object
  domain.BackendCapability:
    enum:
    - deploy
    - delete
    - status
    - sync
    - domains
    - secrets
    - persistent_storage
    - replicas
    - dry_run
    type: string
    x-enum-varnames:
    - BackendCapabilityDeploy
    - BackendCapabilityDelete
    - BackendCapabilityStatus
    - BackendCapabilitySync
    - BackendCapabilityDomains
    - BackendCapabilitySecrets
    - BackendCapabilityPersistentStorage
    - BackendCapabilityReplicas
    - BackendCapabilityDryRun
  domain.BackendRef:
    properties:
      deployment:
        type: string
      ingress_name:
        type: string
      namespace:
        type: string
      service:
        type: string
    type: object
  domain.DeadLetterJob:
    properties:
      failed_at:
        type: string
      final_attempt:
        type: integer
      id:
        type: string
      job:
        $ref: '#/definitions/domain.Job'
      reason:
        type: string
    type: object
  domain.Deployment:
    properties:
      app_id:
        type: string
      backend:
        type: string
      backend_ref:
        $ref: '#/definitions/domain.BackendRef'
      config_snapshot:
        $ref: '#/definitions/domain.AppConfig'
      created_at:
        type: string
      finished_at:
        type: string
      id:
        type: string
      image_ref:
        type: string
      repair_attempts:
        type: integer
      requested_by:
        type: string
      started_at:
        type: string
      status:
        $ref: '#/definitions/domain.DeploymentStatus'
      status_reason:
        type: string
      tenant_id:
        type: string
      updated_at:
        type: string
      version:
        type: integer
    type: object
...
  domain.DeploymentStatus:
    enum:
    - requested
    - queued
    - provisioning
    - running
    - degraded
    - recovering
    - failed
    - cancelled
    - deleting
    - deleted
    type: string
    x-enum-varnames:
    - DeploymentStatusRequested
    - DeploymentStatusQueued
    - DeploymentStatusProvisioning
    - DeploymentStatusRunning
    - DeploymentStatusDegraded
    - DeploymentStatusRecovering
    - DeploymentStatusFailed
    - DeploymentStatusCancelled
    - DeploymentStatusDeleting
    - DeploymentStatusDeleted
  domain.InferenceFallbackProvider:
    properties:
      base_url:
        type: string
      model:
        type: string
      name:
        type: string
      priority:
        type: integer
      timeout_seconds:
        type: integer
    type: object
  domain.Job:
    properties:
      app_id:
        type: string
      attempts:
        type: integer
      available_at:
        type: string
      created_at:
        type: string
      deployment_id:
        type: string
      id:
        type: string
      last_error:
        type: string
      leased_at:
        type: string
      tenant_id:
        type: string
      type:
        $ref: '#/definitions/domain.JobType'
    type: object
  domain.JobType:
    enum:
    - deployment.create
    - deployment.delete
    - deployment.sync
    type: string
    x-enum-varnames:
    - JobTypeDeploymentCreate
    - JobTypeDeploymentDelete
    - JobTypeDeploymentSync
  domain.OpenClawConfig:
    properties:
      agents_markdown:
        type: string
      enabled:
        type: boolean
      existing_secret_name:
        type: string
      extra_env:
        additionalProperties:
          type: string
        type: object
      gateway_bind_address:
        type: string
      gateway_port:
        type: integer
      gateway_token:
        type: string
      provider_api_keys:
        additionalProperties:
          type: string
        type: object
      settings_json:
        type: string
      workspace_storage:
        type: string
    type: object
...
  domain.QueueStats:
    properties:
      dead_letters:
        type: integer
      delayed:
        type: integer
      processing:
        type: integer
      ready:
        type: integer
    type: object
  domain.ReliabilityPolicy:
    properties:
      auto_repair:
        type: boolean
      fallback_providers:
        items:
          $ref: '#/definitions/domain.InferenceFallbackProvider'
        type: array
      health_check_path:
        type: string
      health_check_timeout_seconds:
        type: integer
      max_repair_attempts:
        type: integer
      rollback_on_failed_healthcheck:
        type: boolean
    type: object
  domain.Role:
    enum:
    - owner
    - admin
    - member
    type: string
    x-enum-varnames:
    - RoleOwner
    - RoleAdmin
    - RoleMember
  http.APIKeyCreateRequest:
    properties:
      name:
        type: string
    type: object
  http.APIKeyCreateResultResponse:
    properties:
      key:
        $ref: '#/definitions/domain.APIKey'
      secret:
        type: string
    type: object
  http.APIKeyListResponse:
    properties:
      items:
        items:
          $ref: '#/definitions/domain.APIKey'
        type: array
    type: object
  http.AdminBackendListResponse:
    properties:
      items:
        items:
          $ref: '#/definitions/domain.BackendCapabilities'
        type: array
    type: object
  http.AdminInstanceListResponse:
    properties:
      items:
        items:
          $ref: '#/definitions/domain.AdminInstance'
        type: array
    type: object
  http.AdminPreflightCheck:
    properties:
      id:
        type: string
      label:
        type: string
      message:
        type: string
      severity:
        type: string
      status:
        type: string
    type: object
  http.AdminPreflightResponse:
    properties:
      backend:
        type: string
      capabilities:
        $ref: '#/definitions/domain.BackendCapabilities'
      checks:
        items:
          $ref: '#/definitions/http.AdminPreflightCheck'
        type: array
      default_ingress_host:
        type: string
      development_auth:
        type: boolean
      readiness:
        additionalProperties:
          type: string
        type: object
      recommendations:
        items:
          type: string
        type: array
      repository_driver:
        type: string
      status:
        type: string
    type: object
  http.AppListResponse:
    properties:
      items:
        items:
          $ref: '#/definitions/domain.App'
        type: array
    type: object
  http.AuditLogListResponse:
    properties:
      items:
        items:
          $ref: '#/definitions/domain.AuditLog'
        type: array
    type: object
...
  http.DeadLetterListResponse:
    properties:
      items:
        items:
          $ref: '#/definitions/domain.DeadLetterJob'
        type: array
    type: object
...
  http.ErrorResponse:
    properties:
      code:
        type: string
      message:
        type: string
    type: object
...
  http.LoginURLResponse:
    properties:
      url:
        type: string
    type: object
...
  http.adminCreateUserInput:
    properties:
      display_name:
        type: string
      email:
        type: string
    type: object
  http.adminDeployOpenClawInput:
    properties:
      agents_markdown:
        type: string
      app_name:
        type: string
      app_slug:
        type: string
      domain:
        type: string
      existing_secret_name:
        type: string
      extra_env:
        additionalProperties:
          type: string
        type: object
      gateway_bind_address:
        type: string
      gateway_port:
        type: integer
      gateway_token:
        type: string
      image:
        type: string
      provider_api_keys:
        additionalProperties:
          type: string
        type: object
      public:
        type: boolean
      replicas:
        type: integer
      settings_json:
        type: string
      user_email:
        type: string
      user_name:
        type: string
      workspace_storage:
        type: string
    type: object
  http.adminDeployOpenClawResponse:
    properties:
      app:
        $ref: '#/definitions/domain.App'
      deployment:
        $ref: '#/definitions/domain.Deployment'
      used_existing_app:
        type: boolean
      user:
        $ref: '#/definitions/domain.Actor'
    type: object
  services.CreateAppInput:
    properties:
      config:
        $ref: '#/definitions/domain.AppConfig'
      name:
        type: string
      slug:
        type: string
    type: object
  services.ReconcileResult:
    properties:
      scanned:
        type: integer
      scheduled:
        type: integer
      skipped:
        type: integer
    type: object
  services.ReliabilityStatus:
    properties:
      queue:
        $ref: '#/definitions/domain.QueueStats'
    type: object
...
info:
  contact: {}
  description: |-
    Code-first Swagger docs for the current Clawflux HTTP API.

    Auth behavior:
    - Most /v1 endpoints accept either X-API-Key or X-User-Email.
    - DEVELOPMENT_AUTH=true allows unauthenticated local requests as developer@local.
    - Platform admin endpoints additionally require X-Platform-Admin: true.
  title: Clawflux API
  version: 0.1.0
paths:
...
  /v1/admin/audit-logs:
    get:
      parameters:
      - description: Maximum number of logs to return
        in: query
        maximum: 200
        minimum: 1
        name: limit
        type: integer
      - description: Tenant API key
        in: header
        name: X-API-Key
        type: string
      - description: Development/local user email
        in: header
        name: X-User-Email
        type: string
      - description: Display name
        in: header
        name: X-User-Name
        type: string
      - description: Set to true for platform admin requests
        in: header
        name: X-Platform-Admin
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/http.AuditLogListResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/http.ErrorResponse'
      summary: List tenant audit logs
      tags:
      - Audit
  /v1/admin/backends:
    get:
      parameters:
      - description: Admin email
        in: header
        name: X-User-Email
        required: true
        type: string
      - description: Set to true
        in: header
        name: X-Platform-Admin
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/http.AdminBackendListResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/http.ErrorResponse'
      summary: List configured deployment backends and capabilities
      tags:
      - Admin
  /v1/admin/dead-letters:
    get:
      parameters:
      - description: Admin email
        in: header
        name: X-User-Email
        required: true
        type: string
      - description: Set to true
        in: header
        name: X-Platform-Admin
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/http.DeadLetterListResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/http.ErrorResponse'
      summary: List dead-letter jobs
      tags:
      - Admin
...
  /v1/admin/instances:
    get:
      parameters:
      - description: Admin email
        in: header
        name: X-User-Email
        required: true
        type: string
      - description: Set to true
        in: header
        name: X-Platform-Admin
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/http.AdminInstanceListResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/http.ErrorResponse'
      summary: List all deployed OpenClaw instances across all tenants
      tags:
      - Admin
  /v1/admin/openclaw/deploy:
    post:
      consumes:
      - application/json
      parameters:
      - description: Admin email
        in: header
        name: X-User-Email
        required: true
        type: string
      - description: Set to true
        in: header
        name: X-Platform-Admin
        required: true
        type: string
      - description: OpenClaw deployment payload
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/http.adminDeployOpenClawInput'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/http.adminDeployOpenClawResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/http.ErrorResponse'
      summary: Deploy OpenClaw for a target user
      tags:
      - Admin
  /v1/admin/preflight:
    get:
      parameters:
      - description: Admin email
        in: header
        name: X-User-Email
        required: true
        type: string
      - description: Set to true
        in: header
        name: X-Platform-Admin
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/http.AdminPreflightResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/http.ErrorResponse'
      summary: Inspect runtime readiness and setup guidance
      tags:
      - Admin
  /v1/admin/reconcile:
    post:
      parameters:
      - description: Admin email
        in: header
        name: X-User-Email
        required: true
        type: string
      - description: Set to true
        in: header
        name: X-Platform-Admin
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/services.ReconcileResult'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/http.ErrorResponse'
      summary: Queue reconciliation work for active deployments
      tags:
      - Admin
  /v1/admin/reliability:
    get:
      parameters:
      - description: Admin email
        in: header
        name: X-User-Email
        required: true
        type: string
      - description: Set to true
        in: header
        name: X-Platform-Admin
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/services.ReliabilityStatus'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/http.ErrorResponse'
      summary: Get reliability queue status
      tags:
      - Admin
  /v1/admin/summary:
    get:
      parameters:
      - description: Tenant API key
        in: header
        name: X-API-Key
        type: string
      - description: Development/local user email
        in: header
        name: X-User-Email
        type: string
      - description: Display name
        in: header
        name: X-User-Name
        type: string
      - description: Set to true for platform admin requests
        in: header
        name: X-Platform-Admin
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.AdminSummary'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/http.ErrorResponse'
      summary: Get platform summary
      tags:
      - Admin
  /v1/admin/users:
    post:
      consumes:
      - application/json
      parameters:
      - description: Admin email
        in: header
        name: X-User-Email
        required: true
        type: string
      - description: Set to true
        in: header
        name: X-Platform-Admin
        required: true
        type: string
      - description: User provision payload
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/http.adminCreateUserInput'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/domain.Actor'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/http.ErrorResponse'
      summary: Provision a user as platform admin
      tags:
      - Admin
  /v1/api-keys:
    get:
      parameters:
      - description: Tenant API key
        in: header
        name: X-API-Key
        type: string
      - description: Development/local user email
        in: header
        name: X-User-Email
        type: string
      - description: Display name
        in: header
        name: X-User-Name
        type: string
      - description: Set to true for platform admin requests
        in: header
        name: X-Platform-Admin
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/http.APIKeyListResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/http.ErrorResponse'
      summary: List tenant API keys
      tags:
      - API Keys
    post:
      consumes:
      - application/json
      parameters:
      - description: Tenant API key
        in: header
        name: X-API-Key
        type: string
      - description: Development/local user email
        in: header
        name: X-User-Email
        type: string
      - description: Display name
        in: header
        name: X-User-Name
        type: string
      - description: Set to true for platform admin requests
        in: header
        name: X-Platform-Admin
        type: string
      - description: API key create request
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/http.APIKeyCreateRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/http.APIKeyCreateResultResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/http.ErrorResponse'
      summary: Create a tenant API key
      tags:
      - API Keys
...
  /v1/apps:
    get:
      parameters:
      - description: Tenant API key
        in: header
        name: X-API-Key
        type: string
      - description: Development/local user email
        in: header
        name: X-User-Email
        type: string
      - description: Display name
        in: header
        name: X-User-Name
        type: string
      - description: Set to true for platform admin requests
        in: header
        name: X-Platform-Admin
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/http.AppListResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/http.ErrorResponse'
      summary: List apps for the current tenant
      tags:
      - Apps
    post:
      consumes:
      - application/json
      parameters:
      - description: Tenant API key
        in: header
        name: X-API-Key
        type: string
      - description: Development/local user email
        in: header
        name: X-User-Email
        type: string
      - description: Display name
        in: header
        name: X-User-Name
        type: string
      - description: Set to true for platform admin requests
        in: header
        name: X-Platform-Admin
        type: string
      - description: App definition
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/services.CreateAppInput'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/domain.App'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/http.ErrorResponse'
      summary: Create an app
      tags:
      - Apps
...
  /v1/auth/medium/callback:
    get:
      parameters:
      - description: OAuth callback code
        in: query
        name: code
        type: string
      - description: Redirect URI
        in: query
        name: redirect_uri
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Actor'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/http.ErrorResponse'
      summary: Handle Medium OAuth callback
      tags:
      - Auth
  /v1/auth/medium/login:
    get:
      parameters:
      - description: Redirect URI
        in: query
        name: redirect_uri
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/http.LoginURLResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/http.ErrorResponse'
      summary: Get Medium OAuth login URL
      tags:
      - Auth
...
  /v1/me:
    get:
      parameters:
      - description: Tenant API key
        in: header
        name: X-API-Key
        type: string
      - description: Development/local user email
        in: header
        name: X-User-Email
        type: string
      - description: Display name
        in: header
        name: X-User-Name
        type: string
      - description: Set to true for platform admin requests
        in: header
        name: X-Platform-Admin
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Actor'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/http.ErrorResponse'
      summary: Return the current actor
      tags:
      - Auth
schemes:
- http
securityDefinitions:
  APIKeyHeader:
    in: header
    name: X-API-Key
    type: apiKey
  UserEmailHeader:
    in: header
    name: X-User-Email
    type: apiKey
swagger: "2.0"